Authentication

+++ src/main/java/com/novi/DemoDrop/controllers/DemoController.java
@@ -3,12 +3,29 @@ package com.novi.DemoDrop.controllers;
 import com.novi.DemoDrop.Dto.InputDto.DemoInputDto;
 import com.novi.DemoDrop.Dto.InputDto.IdInputDto;
 import com.novi.DemoDrop.Dto.OutputDto.DemoOutputDto;
+import com.novi.DemoDrop.exceptions.BadRequestException;
 import com.novi.DemoDrop.exceptions.RecordNotFoundException;
+import com.novi.DemoDrop.Dto.OutputDto.FileUploadResponse;
+import com.novi.DemoDrop.models.Demo;
+import com.novi.DemoDrop.repositories.DemoRepository;
 import com.novi.DemoDrop.services.DemoService;
+import org.springframework.core.io.Resource;
+import jakarta.servlet.http.HttpServletRequest;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import java.io.IOException;
+import java.util.Collection;
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
 @CrossOrigin
 @RequestMapping("demos")
@@ -16,9 +33,11 @@ import java.util.List;
 public class DemoController {
 
     private final DemoService demoService;
+    private final DemoRepository demoRepository;
 
-    public DemoController(DemoService demoService) {
+    public DemoController(DemoService demoService, DemoRepository demoRepository) {
         this.demoService = demoService;
+        this.demoRepository = demoRepository;
     }
 
     @GetMapping
@@ -34,15 +53,60 @@ public class DemoController {
         return ResponseEntity.ok(demoOutputDto);
     }
 
+    @GetMapping("/mydemos/{djId}")
+    public ResponseEntity<List<DemoOutputDto>> getAllMyDemos(@PathVariable Long djId) {
+        List<DemoOutputDto> demoOutputDtos = demoService.getAllMyDemos(djId);
+        return ResponseEntity.ok(demoOutputDtos);
+
+    }
+
+    @GetMapping("/{demoId}/download")
+    public ResponseEntity<Resource> downloadMp3File(@PathVariable Long demoId, HttpServletRequest request) {
+        Resource resource = demoService.downloadFile(demoId);
+        String mimeType;
+
+        try {
+            mimeType = request.getServletContext().getMimeType(resource.getFile().getAbsolutePath());
+        } catch (IOException e) {
+            mimeType = MediaType.APPLICATION_OCTET_STREAM_VALUE;
+        }
+        return ResponseEntity.ok().contentType(MediaType.parseMediaType(mimeType)).header(HttpHeaders.CONTENT_DISPOSITION, "inline;fileName=" + resource.getFilename()).body(resource);
+    }
+
     @PostMapping
     public ResponseEntity<Object> createDemo(@RequestBody DemoInputDto demoInputDto) {
         DemoOutputDto demoOutputDto = demoService.createDemo(demoInputDto);
         return ResponseEntity.ok(demoOutputDto);
     }
 
+    @PostMapping("/mp3file")
+    public FileUploadResponse mp3FileUpload(@RequestParam("file") MultipartFile file, @RequestParam Long demoId) {
+        Optional<Demo> demoOptional = demoRepository.findById(demoId);
+        if (demoOptional.isPresent()) {
+            Demo d = demoOptional.get();
+            String uri = ServletUriComponentsBuilder.fromCurrentContextPath().path("/download").path(Objects.requireNonNull(file.getOriginalFilename())).toUriString();
+
+            String contentType = file.getContentType();
+
+            String fileName = demoService.storeMP3File(file, demoId);
+
+            return new FileUploadResponse(fileName, contentType, uri);
+
+        } else {
+            throw new RecordNotFoundException("No demo found to add MP3 to");
+        }
+
+    }
+
     @DeleteMapping("/{id}")
-    public ResponseEntity<Object> removeDemo (@PathVariable Long id) {
-        boolean isDeleted = demoService.deleteDemo(id);
+    public ResponseEntity<Object> removeDemo(@PathVariable Long id) {
+
+        Optional<Demo> demoOptional = demoRepository.findById(id);
+        boolean isDeleted = false;
+        if (demoOptional.isPresent()) {
+            Demo d = demoOptional.get();
+                isDeleted = demoService.deleteDemo(id);
+        }
         if (isDeleted) {
             return ResponseEntity.ok().body("Element is deleted");
         } else {
@@ -50,15 +114,4 @@ public class DemoController {
         }
 
     }
-    // TO-DO: voeg nog @Valid toe voor @RequestBody als validation dependency geinjecteerd is
-    // this method assigns an existing reply to a demo
-    @PutMapping("/{id}/reply-to-demo")
-    public void assignRemoteControllerToTelevision (@PathVariable("id") Long id, @RequestBody IdInputDto input) {
-        demoService.assignReplyToDemo(id, input.id);
-    }
-
-
-    // Nog toe te voegen requests:
-    // update demo request (maar echt?)
-    // get demo info (id, artistname and songname and email - voor de reactie) Maar misschien moet deze in de ReplyToDemoKlasse?
 }
